Administrative units of the Indian subcontinent

Names and codes of administrative units (three levels) used by eBird in their datasets for the eight countries of the Indian subcontinent.

There is often a mismatch in information on administrative units in different sources, so this data can be useful for any work involving eBird data, particularly in cases with low data where the downloaded dataset might not have observations from all relevant units or subunits.

This data is obtained from the API and should therefore stay up to date with each package release. However, it may be outdated if changes occur before the package is updated. In such cases, one can use rebird::ebirdsubregionlist() manually (section See Also).

Usage

ebird_admin_units

Format

A tibble with 1,135 observations and 6 variables (names and 2-letter ISO codes).

See also

Examples

  ebird_admin_units
#> # A tibble: 1,135 × 6
#>    COUNTRY.CODE COUNTRY STATE.CODE STATE                      COUNTY.CODE COUNTY
#>    <chr>        <chr>   <chr>      <chr>                      <chr>       <chr> 
#>  1 IN           India   IN-AN      Andaman and Nicobar Islan… IN-AN-AN    North…
#>  2 IN           India   IN-AN      Andaman and Nicobar Islan… IN-AN-AS    South…
#>  3 IN           India   IN-AN      Andaman and Nicobar Islan… IN-AN-NI    Nicob…
#>  4 IN           India   IN-AP      Andhra Pradesh             IN-AP-AN    Anant…
#>  5 IN           India   IN-AP      Andhra Pradesh             IN-AP-CH    Chitt…
#>  6 IN           India   IN-AP      Andhra Pradesh             IN-AP-CU    YSR D…
#>  7 IN           India   IN-AP      Andhra Pradesh             IN-AP-EG    East …
#>  8 IN           India   IN-AP      Andhra Pradesh             IN-AP-GU    Guntur
#>  9 IN           India   IN-AP      Andhra Pradesh             IN-AP-KR    Krish…
#> 10 IN           India   IN-AP      Andhra Pradesh             IN-AP-KU    Kurno…
#> # ℹ 1,125 more rows
  # get data for only India
  dplyr::filter(ebird_admin_units, COUNTRY.CODE == "IN")
#> # A tibble: 735 × 6
#>    COUNTRY.CODE COUNTRY STATE.CODE STATE                      COUNTY.CODE COUNTY
#>    <chr>        <chr>   <chr>      <chr>                      <chr>       <chr> 
#>  1 IN           India   IN-AN      Andaman and Nicobar Islan… IN-AN-AN    North…
#>  2 IN           India   IN-AN      Andaman and Nicobar Islan… IN-AN-AS    South…
#>  3 IN           India   IN-AN      Andaman and Nicobar Islan… IN-AN-NI    Nicob…
#>  4 IN           India   IN-AP      Andhra Pradesh             IN-AP-AN    Anant…
#>  5 IN           India   IN-AP      Andhra Pradesh             IN-AP-CH    Chitt…
#>  6 IN           India   IN-AP      Andhra Pradesh             IN-AP-CU    YSR D…
#>  7 IN           India   IN-AP      Andhra Pradesh             IN-AP-EG    East …
#>  8 IN           India   IN-AP      Andhra Pradesh             IN-AP-GU    Guntur
#>  9 IN           India   IN-AP      Andhra Pradesh             IN-AP-KR    Krish…
#> 10 IN           India   IN-AP      Andhra Pradesh             IN-AP-KU    Kurno…
#> # ℹ 725 more rows