Filter eBird data for eligible checklists

For various usecases, it is required to filter the eBird dataset and use only "eligible" checklists. An example is the monthly eBird challenges run by various collectives across the world. As such, definitions of "eligibility" tend to vary, but there are two popularly used definitions which can be accessed via the standard argument. On the other hand, it is also possible to create a custom eligibility filter, by setting standard = NULL and providing input to the other arguments.

Usage

filter_eligible(
  data,
  standard = "bci",
  complete = NULL,
  duration = NULL,
  counts = NULL
)

Arguments

data

eBird data object

standard

character; a standard eligibility criterion (default "bci", or "ebird")

complete

logical; whether or not to filter for complete checklists (default NULL)

duration

numerical; whether or not to filter for a specified checklist duration (>=; default NULL)

counts

logical; whether or not to filter for true counts (i.e., remove lists reporting Xs; default NULL)

Value

An updated eBird data object filtered as specified

See also